26thOct2020

+++ apps/CollectSweepResults.py
@@ -5,25 +5,28 @@ import shutil
 import os
 import string
 import math
+import os.path
+from os import path
 
 
 
 if __name__=="__main__":
     
-    NewFolder = "/data/vascrem/testoutput/ParameterSweep/Cylinder/Parameteres2/CollectedResults/"
+    NewFolder = "/data/vascrem/testoutput/Tissue2dVirusInfection_InfectiousPercent_S/CollectedResults/"
+    OldFolder = "/data/vascrem/testoutput/Tissue2dVirusInfection_InfectiousPercent_S/"
+    # "InitialInfec_0.05_Diffusion_0.5/6/results_from_time_0/pde_results_ext_virus_4000.vtu
     os.mkdir(NewFolder)
-    AreaParameter = [6, 6.5, 7,7.5, 8]
-    DilationParameter =[6, 6.5, 7,7.5, 8]
-    DeformationParamter = [6, 6.5, 7,7.5, 8]
+    Parameter1 = [0.05,0.1,0.01]
+    Parameter2 =[1,0.1,0.5,0.01,0.05,0.005]
+    Parameter3 = [1,2,3,4,5,6,7,8,9, 10]
 
-    for i in AreaParameter:
-        for j in DilationParameter:
-            for k in DeformationParamter:
-                Oldfile = "/data/vascrem/testoutput/ParameterSweep/Cylinder/Parameteres2/Param_"+  str(i) +"_DilationParam_"+str(j) + "_DeformationParam_" +str(k) +"/results_from_time_30/results.viznodes" 
-
-                NewFile = NewFolder+"Area_"+  str(i) +"_Dil_"+str(j) + "_Def_" +str(k)+".viznodes"
-                # print"Hit"
-                shutil.copy(Oldfile, NewFile)
+    for i in Parameter1:
+        for j in Parameter2:
+            for k in Parameter3:
+                Oldfile = OldFolder + "InitialInfec_" + str(i) + "_Diffusion_" + str(j) + "/" + str(k) + "/results_from_time_0/viruscelltypes.dat" 
+                if path.exists(Oldfile):
+                    NewFile = NewFolder + "InitialInfec_" + str(i) + "_Diffusion_" + str(j) + "_" + str(k)+"_viruscelltypes.dat" 
+                    shutil.copy(Oldfile, NewFile)
    
     print '\n ********* ------ Finished ------ ********* \n'
